+++ strings/json_normalize.c
@@ -264,9 +264,9 @@ json_norm_array_append_value(struct json_norm_array *arr,
 int
 json_norm_init_dynamic_array(size_t element_size, void *where)
 {
-  const uint init_alloc= 20;
-  const uint alloc_increment= 20;
-  return my_init_dynamic_array(PSI_JSON, where, (uint)element_size,
+  const size_t init_alloc= 20;
+  const size_t alloc_increment= 20;
+  return my_init_dynamic_array(PSI_JSON, where, element_size,
                                init_alloc, alloc_increment,
                                JSON_MALLOC_FLAGS);
 }
